src/utility_functions/requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
transformers>=4.20.0
Pillow>=9.0.0
qrcode>=7.3.0
cryptography>=40.0.0
reportlab>=3.6.0
